Filters are working :o
but it turns out that the open file method is not.  it’s because i
changed adding a string array with values set to the data stored, but
it was probably adding that same array, so only one actor’s data would
be opened in the file. i’m going to work on the addition of multiple
new string arrays… right now i have an error out of bounds exception,
but i think i know why…