i tried
besides some minor things like… crashing when a string is entered while
expecting an int… the opening of a saved file is the only major thing
that didn’t work as i intended it to… funny thing is that that was one
of the first things i coded, and even planned out three times on paper.
writing to the file in the translated text works just fine, but i
cannot read from the file without my app crashing.  this was the only
thing i worked on for the past two hours… i got close to printing the
correct things, but i do not know if the correct things were in the
arrays (i think not…). at the least, the search filters work great, and
actors can be added and deleted and saved just fine—they just can’t be
retrieved.